Fix UTM to LL overflow on lambda0 calc. Since zone is a u32, the subsequent math was done as u32, and could go below 0. Casting zone to an i32

+++ src/geo.rs
@@ -320,7 +320,7 @@ pub fn utm2ll(utm: &utm_grid::utm_grid, ellipsoid: &geo_ellipsoid::geo_ellipsoid
     let phi = taui.atan();
     let mut lambda = sinetap.atan2(cosxip);
 
-    let lambda0 = (((z - 1) * 6 - 180 + 3) as f64).to_radians();
+    let lambda0 = (((z as i32 - 1) * 6 - 180 + 3) as f64).to_radians();
     lambda += lambda0;
 
     ret_val.x = phi;
